+++ src/app/anhoriga/components/TipCard.tsx
