+++ src/app/anhoriga/components/TipsCards.tsx
