+++ src/app/anhoriga/components/WhatIsBipolar.tsx
@@ -1,4 +1,4 @@
-import SectionTitle from "../components/shared/headings/SectionTitle";
+import SectionTitle from "../../components/shared/headings/SectionTitle";
 import { FaQuestionCircle } from "react-icons/fa";
 
 const WhatIsBipolar = () => {
